+++ projects/angular-split/src/lib/component/split.component.ts
@@ -12,7 +12,6 @@ import {
   ViewChildren,
   QueryList,
   EventEmitter,
-  ViewEncapsulation,
   Inject,
   Optional,
   ContentChild,
@@ -110,7 +109,7 @@ import { SplitGutterDirective } from '../gutter/split-gutter.directive'
         [attr.aria-orientation]="direction"
         [attr.aria-valuemin]="area.minSize"
         [attr.aria-valuemax]="area.maxSize"
-        [attr.aria-valuenow]="area.size"
+        [attr.aria-valuenow]="area.size === '*' ? null : area.size"
         [attr.aria-valuetext]="getAriaAreaSizeText(area.size)"
       >
         <ng-container *ngIf="customGutter?.template; else defaultGutterTpl">
@@ -306,12 +305,12 @@ export class SplitComponent implements AfterViewInit, OnDestroy {
   private isDragging = false
   private isWaitingClear = false
   private isWaitingInitialMove = false
-  private dragListeners: Array<Function> = []
+  private dragListeners: Array<() => void> = []
   private snapshot: ISplitSnapshot | null = null
   private startPoint: IPoint | null = null
   private endPoint: IPoint | null = null
 
-  public readonly displayedAreas: Array<IArea> = []
+  readonly displayedAreas: Array<IArea> = []
   private readonly hiddenAreas: Array<IArea> = []
 
   @ViewChildren('gutterEls') private gutterEls: QueryList<ElementRef>
@@ -319,7 +318,7 @@ export class SplitComponent implements AfterViewInit, OnDestroy {
   _clickTimeout: number | null = null
   draggedGutterNum: number = undefined
 
-  public ngAfterViewInit() {
+  ngAfterViewInit() {
     this.ngZone.runOutsideAngular(() => {
       // To avoid transition at first rendering
       setTimeout(() => this.renderer.addClass(this.elRef.nativeElement, 'as-init'))
@@ -330,7 +329,7 @@ export class SplitComponent implements AfterViewInit, OnDestroy {
     return this.displayedAreas.length === 0 ? 0 : this.displayedAreas.length - 1
   }
 
-  public addArea(component: SplitAreaDirective): void {
+  addArea(component: SplitAreaDirective): void {
     const newArea: IArea = {
       component,
       order: 0,
@@ -350,7 +349,7 @@ export class SplitComponent implements AfterViewInit, OnDestroy {
     }
   }
 
-  public removeArea(component: SplitAreaDirective): void {
+  removeArea(component: SplitAreaDirective): void {
     if (this.displayedAreas.some((a) => a.component === component)) {
       const area = this.displayedAreas.find((a) => a.component === component)
       this.displayedAreas.splice(this.displayedAreas.indexOf(area), 1)
@@ -362,13 +361,13 @@ export class SplitComponent implements AfterViewInit, OnDestroy {
     }
   }
 
-  public updateArea(component: SplitAreaDirective, resetOrders: boolean, resetSizes: boolean): void {
+  updateArea(component: SplitAreaDirective, resetOrders: boolean, resetSizes: boolean): void {
     if (component.visible === true) {
       this.build(resetOrders, resetSizes)
     }
   }
 
-  public showArea(component: SplitAreaDirective): void {
+  showArea(component: SplitAreaDirective): void {
     const area = this.hiddenAreas.find((a) => a.component === component)
     if (area === undefined) {
       return
@@ -380,7 +379,7 @@ export class SplitComponent implements AfterViewInit, OnDestroy {
     this.build(true, true)
   }
 
-  public hideArea(comp: SplitAreaDirective): void {
+  hideArea(comp: SplitAreaDirective): void {
     const area = this.displayedAreas.find((a) => a.component === comp)
     if (area === undefined) {
       return
@@ -396,11 +395,11 @@ export class SplitComponent implements AfterViewInit, OnDestroy {
     this.build(true, true)
   }
 
-  public getVisibleAreaSizes(): IOutputAreaSizes {
+  getVisibleAreaSizes(): IOutputAreaSizes {
     return this.displayedAreas.map((a) => a.size)
   }
 
-  public setVisibleAreaSizes(sizes: IOutputAreaSizes): boolean {
+  setVisibleAreaSizes(sizes: IOutputAreaSizes): boolean {
     if (sizes.length !== this.displayedAreas.length) {
       return false
     }
@@ -412,8 +411,8 @@ export class SplitComponent implements AfterViewInit, OnDestroy {
       return false
     }
 
-    // @ts-ignore
-    this.displayedAreas.forEach((area, i) => (area.component._size = formattedSizes[i]))
+    // @@ts-expect-error
+    this.displayedAreas.forEach((area, i) => (area.component.size = formattedSizes[i]))
 
     this.build(false, true)
     return true
@@ -470,7 +469,7 @@ export class SplitComponent implements AfterViewInit, OnDestroy {
             if (wildcardSizeAreas.length === 0 && this.displayedAreas.length > 0) {
               this.displayedAreas.forEach((area, i) => {
                 area.size = i === 0 ? '*' : area.component.size
-                area.minSize = i === 0 ? null : getAreaMinSize(area)
+                area.minSize = i === 0 ? area.component.minSize : getAreaMinSize(area)
                 area.maxSize = i === 0 ? null : getAreaMaxSize(area)
               })
             } else if (wildcardSizeAreas.length > 1) {
@@ -566,7 +565,7 @@ export class SplitComponent implements AfterViewInit, OnDestroy {
     }
   }
 
-  public clickGutter(event: MouseEvent | TouchEvent, gutterNum: number): void {
+  clickGutter(event: MouseEvent | TouchEvent, gutterNum: number): void {
     const tempPoint = getPointFromEvent(event)
 
     // Be sure mouseup/touchend happened if touch/cursor is not moved.
@@ -592,7 +591,7 @@ export class SplitComponent implements AfterViewInit, OnDestroy {
     }
   }
 
-  public startKeyboardDrag(event: KeyboardEvent, gutterOrder: number, gutterNum: number) {
+  startKeyboardDrag(event: KeyboardEvent, gutterOrder: number, gutterNum: number) {
     if (this.disabled === true || this.isWaitingClear === true) {
       return
     }
@@ -613,7 +612,7 @@ export class SplitComponent implements AfterViewInit, OnDestroy {
     this.stopDragging()
   }
 
-  public startMouseDrag(event: MouseEvent | TouchEvent, gutterOrder: number, gutterNum: number): void {
+  startMouseDrag(event: MouseEvent | TouchEvent, gutterOrder: number, gutterNum: number): void {
     if (this.customGutter && !this.customGutter.canStartDragging(event.target as HTMLElement, gutterNum)) {
       return
     }
@@ -789,7 +788,9 @@ export class SplitComponent implements AfterViewInit, OnDestroy {
 
     // Each gutter side areas can't absorb all offset
     if (areasBefore.remain !== 0 && areasAfter.remain !== 0) {
+      // TODO: fix this emty block
       if (Math.abs(areasBefore.remain) === Math.abs(areasAfter.remain)) {
+        /* empty */
       } else if (Math.abs(areasBefore.remain) > Math.abs(areasAfter.remain)) {
         areasAfter = getGutterSideAbsorptionCapacity(
           this.unit,
@@ -827,7 +828,7 @@ export class SplitComponent implements AfterViewInit, OnDestroy {
       // Hack because of browser messing up with sizes using calc(X% - Ypx) -> el.getBoundingClientRect()
       // If not there, playing with gutters makes total going down to 99.99875% then 99.99286%, 99.98986%,..
       const all = [...areasBefore.list, ...areasAfter.list]
-      const wildcardArea = all.find((a) => a.percentAfterAbsorption == '*')
+      const wildcardArea = all.find((a) => a.percentAfterAbsorption === '*')
       // In case we have a wildcard area - always align the percents on the wildcard area.
       const areaToReset =
         wildcardArea ??
@@ -899,7 +900,7 @@ export class SplitComponent implements AfterViewInit, OnDestroy {
     })
   }
 
-  public notify(type: 'start' | 'progress' | 'end' | 'click' | 'dblclick' | 'transitionEnd', gutterNum: number): void {
+  notify(type: 'start' | 'progress' | 'end' | 'click' | 'dblclick' | 'transitionEnd', gutterNum: number): void {
     const sizes = this.getVisibleAreaSizes()
 
     if (type === 'start') {
@@ -920,11 +921,11 @@ export class SplitComponent implements AfterViewInit, OnDestroy {
     }
   }
 
-  public ngOnDestroy(): void {
+  ngOnDestroy(): void {
     this.stopDragging()
   }
 
-  public collapseArea(comp: SplitAreaDirective, newSize: number, gutter: 'left' | 'right'): void {
+  collapseArea(comp: SplitAreaDirective, newSize: number, gutter: 'left' | 'right'): void {
     const area = this.displayedAreas.find((a) => a.component === comp)
     if (area === undefined) {
       return
@@ -942,7 +943,7 @@ export class SplitComponent implements AfterViewInit, OnDestroy {
     this.updateArea(comp, false, false)
   }
 
-  public expandArea(comp: SplitAreaDirective): void {
+  expandArea(comp: SplitAreaDirective): void {
     const area = this.displayedAreas.find((a) => a.component === comp)
     if (area === undefined) {
       return
@@ -959,7 +960,7 @@ export class SplitComponent implements AfterViewInit, OnDestroy {
     this.updateArea(comp, false, false)
   }
 
-  public getAriaAreaSizeText(size: IAreaSize): string {
+  getAriaAreaSizeText(size: IAreaSize): string {
     if (size === '*') {
       return null
     }