+++ src/app/examples/dir-rtl/dir-rtl.component.ts
@@ -1,13 +1,10 @@
-import { Component, ChangeDetectionStrategy } from '@angular/core'
+import { Component, ChangeDetectionStrategy, HostBinding } from '@angular/core'
 import { AComponent } from '../../ui/components/AComponent'
 import { ISplitDir, ISplitDirection } from 'angular-split'
 
 @Component({
-  selector: 'sp-ex-dir_rtl',
+  selector: 'sp-ex-dir-rtl',
   changeDetection: ChangeDetectionStrategy.OnPush,
-  host: {
-    class: 'split-example-page',
-  },
   styles: [
     `
       .btns {
@@ -74,6 +71,8 @@ import { ISplitDir, ISplitDirection } from 'angular-split'
   `,
 })
 export class DirRtlComponent extends AComponent {
+  @HostBinding('class') class = 'split-example-page'
+
   dir: ISplitDir = 'rtl'
   direction: ISplitDirection = 'horizontal'
 }