+++ src/app/examples/nested-split/nested-split.component.ts
@@ -1,13 +1,10 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core'
+import { ChangeDetectionStrategy, Component, HostBinding } from '@angular/core'
 
 import { AComponent } from '../../ui/components/AComponent'
 
 @Component({
   selector: 'sp-ex-nested',
   changeDetection: ChangeDetectionStrategy.OnPush,
-  host: {
-    class: 'split-example-page',
-  },
   template: `
     {{ testChangeDetectorRun() }}
     <div class="container">
@@ -72,4 +69,6 @@ import { AComponent } from '../../ui/components/AComponent'
     </div>
   `,
 })
-export class NestedComponent extends AComponent {}
+export class NestedComponent extends AComponent {
+  @HostBinding('class') class = 'split-example-page'
+}